requirements.txt
openai>=1.0.0
semantic-kernel>=0.91.7
python-dotenv>=1.0.0
gitpython
requests